chore: moving test only used functions into test
